Relax guarantees around FileHandle's state
Because `FileHandle` is currently implemented as a struct and is
implicitly copyable, we can't make strong guarantees about its state.
If the handle is copied and only a single copy is closed, any other
copies won't know. Instead of refactoring this to a class, let's instead
punt and name our internal state "isKnownClosed" and "isKnownInvalid" to
hint that these state changes are inherently caches based on the current
instance.

In the future it should be possible to mark FileHandle as move-only to
prevent this unintended sharing.